termine creacion  de modelos

+++ database/migrations/20230411115051-create-user.js
@@ -28,7 +28,7 @@ module.exports = {
         type: Sequelize.DATE
       },
       phone: {
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(55)
       },
       news: {
         type: Sequelize.INTEGER
